+++ Common/src/main/java/net/potionstudios/biomeswevegone/tags/BWGBlockTags.java
@@ -17,6 +17,15 @@ public class BWGBlockTags {
     public static final TagKey<Block> SNOWY_PLANT_PLACEABLE = create("snowy_plant_placeable");
     public static final TagKey<Block> BWG_MUSHROOM_PLACEABLE = create("bwg_mushroom_placeable");
     public static final TagKey<Block> HYDRANGEA_BUSH_PLACEABLE = create("hydrangea_bush_placeable");
+    public static final TagKey<Block> TALL_ALLIUMS = create("flowers/alliums/tall");
+    public static final TagKey<Block> SHORT_ALLIUMS = create("flowers/alliums/short");
+    public static final TagKey<Block> ALLIUM_FLOWER_BUSHES = create("flowers/alliums/flower_bushes");
+    public static final TagKey<Block> ALLIUMS = create("flowers/alliums");
+    public static final TagKey<Block> ROSES = create("flowers/roses");
+    public static final TagKey<Block> TULIPS = create("flowers/tulips");
+    public static final TagKey<Block> AMARANTH = create("flowers/amaranth");
+    public static final TagKey<Block> SAGES = create("flowers/sages");
+    public static final TagKey<Block> DAFFODILS = create("flowers/daffodils");
 
     private static TagKey<Block> create(String name) {
         return TagKey.create(Registries.BLOCK, BiomesWeveGone.id(name));